defra-native-fs-runner: split lib.rs into protocol/tools/runner modules
