+++ crates/defra-native-fs-runner/src/traversal.rs
@@ -1,0 +1,8 @@
+mod common;
+mod glob_matches;
+mod grep_matches;
+mod list_entries;
+
+pub(crate) use glob_matches::collect_glob_matches;
+pub(crate) use grep_matches::collect_grep_matches;
+pub(crate) use list_entries::collect_entries;
